change all float to double

+++ src/server/exchange.c
@@ -4,14 +4,14 @@
 
 book * buyBook = NULL;
 book * sellBook = NULL;
-float buyPrice = 0;
-float sellPrice = 0;
+double buyPrice = 0;
+double sellPrice = 0;
 
 // not any of orders are freed currently
 // decide when to free once the way to inform client is decided
 /// status of order: 0 not matched, 1 matched, 2 canceled
 
-order * placeBuyOrder(float price, int amount, int orderId, int userId){
+order * placeBuyOrder(double price, int amount, int orderId, int userId){
     order * currentOrder = generateOrder(price, amount, orderId, userId);
     order * temp = matchOrder(currentOrder, &sellBook);
     // if the order is not matched, add it to buy book and update the best sell price
@@ -26,7 +26,7 @@ order * placeBuyOrder(float price, int amount, int orderId, int userId){
             book * tempBook = NULL;
             HASH_FIND(hh, sellBook, &price, 4, tempBook);
             if(tempBook == NULL){
-                float tempBuyPrice = 0;
+                double tempBuyPrice = 0;
                 // loop through the sell book to find the lowest price (best buy price)
                 for(tempBook = sellBook; tempBook != NULL; tempBook = (book *)tempBook -> hh.next){
                     if(tempBuyPrice > tempBook -> price || tempBuyPrice == 0){
@@ -43,7 +43,7 @@ order * placeBuyOrder(float price, int amount, int orderId, int userId){
     return currentOrder;
 }
 
-order * cancelBuyOrder(float price, int exchangeId){
+order * cancelBuyOrder(double price, int exchangeId){
     book * tempBook = NULL;
     priceBucket * tempPriceBucket = NULL;
     priceBucket * orderPriceBucket = NULL;
@@ -66,7 +66,7 @@ order * cancelBuyOrder(float price, int exchangeId){
     return tempOrder;
 }
 
-order * placeSellOrder(float price, int amount, int orderId, int userId){
+order * placeSellOrder(double price, int amount, int orderId, int userId){
     order * currentOrder = generateOrder(price, amount, orderId, userId);
     order * temp = matchOrder(currentOrder, &buyBook);
     // if the order is not matched, add it to sell book and update the best buy price
@@ -81,7 +81,7 @@ order * placeSellOrder(float price, int amount, int orderId, int userId){
             book * tempBook = NULL;
             HASH_FIND(hh, buyBook, &price, 4, tempBook);
             if(tempBook == NULL){
-                float tempSellPrice = 0;
+                double tempSellPrice = 0;
                 // loop through the sell book to find the lowest price (best buy price)
                 for(tempBook = buyBook; tempBook != NULL; tempBook = (book *)tempBook -> hh.next){
                     if(tempSellPrice < tempBook -> price){
@@ -98,7 +98,7 @@ order * placeSellOrder(float price, int amount, int orderId, int userId){
     return currentOrder;
 }
 
-order * cancelSellOrder(float price, int exchangeId){
+order * cancelSellOrder(double price, int exchangeId){
     book * tempBook = NULL;
     priceBucket * tempPriceBucket = NULL;
     priceBucket * orderPriceBucket = NULL;
@@ -122,7 +122,7 @@ order * cancelSellOrder(float price, int exchangeId){
 }
 
 order * matchOrder(order * curOrder, book ** curBook){
-    float price = curOrder -> price;
+    double price = curOrder -> price;
     int amount = curOrder -> amount;
     book * tempBook = NULL;
     order * tempOrder = NULL;
