+++ src/server/order.h
@@ -5,7 +5,7 @@
 // order's status: 0 in book; 1 matched; 2 cancel; 3 NACH
 typedef struct order
 {
-	float price;
+	double price;
 	int amount;
 	int orderId;
 	int exchangeId;
@@ -17,6 +17,6 @@ typedef struct order
 * generate order with a generated key exchangeId and a initial status 0
 * status: 0 not matched; 1 matched; 2 canceled
 */
-order * generateOrder(float price, int amount, int orderId, int userId);
+order * generateOrder(double price, int amount, int orderId, int userId);
 
 #endif /* ORDER_H_ */
